backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "ipykernel>=7.2.0",
    "networkx>=3.0",
    "numpy>=2.4.3",
    "opencv-python-headless>=4.8.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-image>=0.21.0",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]
